+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Do not unlink gradle projects during the project refresh [#956](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/956)
 - Changed output directory for custom modules to `classes` [#958](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/958)
 
+### Other
+- Updated project libraries [#959](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/959)
+
 ### Fixes
 - `AntInstallation` class not found on project refresh [#955](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/955)